requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.12.0
pyyaml>=6.0
python-dotenv>=1.0.0
openai>=1.0.0
google-generativeai>=0.8.0
matplotlib>=3.7.0
